requirements.txt
alembic==1.4.2
Babel==2.8.0
bandit==1.6.2
blinker==1.4
certifi==2020.6.20
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
elasticsearch==7.8.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-BS4==4.5.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gitdb==4.0.5
GitPython==3.1.8
idna==2.10
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
parso==0.7.0
pbr==5.5.0
pwgen==0.8.2.post0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
repoze.lru==0.7
Routes==2.4.1
rq==1.4.3
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.18
stevedore==3.2.1
ujson==3.0.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
